Move getTopProvider to ProviderUtils
