+++ auth/src/main/java/com/firebase/ui/auth/ui/accountlink/WelcomeBackIdpPrompt.java
@@ -30,7 +30,7 @@ import com.firebase.ui.auth.ErrorCodes;
 import com.firebase.ui.auth.IdpResponse;
 import com.firebase.ui.auth.R;
 import com.firebase.ui.auth.ResultCodes;
-import com.firebase.ui.auth.provider.AuthCredentialHelper;
+import com.firebase.ui.auth.provider.ProviderUtils;
 import com.firebase.ui.auth.provider.FacebookProvider;
 import com.firebase.ui.auth.provider.GoogleProvider;
 import com.firebase.ui.auth.provider.IdpProvider;
@@ -76,7 +76,7 @@ public class WelcomeBackIdpPrompt extends AppCompatBase implements IdpCallback {
         setContentView(R.layout.welcome_back_idp_prompt_layout);
 
         IdpResponse newUserIdpResponse = IdpResponse.fromResultIntent(getIntent());
-        mPrevCredential = AuthCredentialHelper.getAuthCredential(newUserIdpResponse);
+        mPrevCredential = ProviderUtils.getAuthCredential(newUserIdpResponse);
 
         User oldUser = User.getUser(getIntent());
 
@@ -140,7 +140,7 @@ public class WelcomeBackIdpPrompt extends AppCompatBase implements IdpCallback {
             return; // do nothing
         }
 
-        AuthCredential newCredential = AuthCredentialHelper.getAuthCredential(idpResponse);
+        AuthCredential newCredential = ProviderUtils.getAuthCredential(idpResponse);
         if (newCredential == null) {
             Log.e(TAG, "No credential returned");
             finish(ResultCodes.CANCELED, IdpResponse.getErrorCodeIntent(ErrorCodes.UNKNOWN_ERROR));
