+++ auth/src/main/java/com/firebase/ui/auth/ui/idp/CredentialSignInHandler.java
@@ -23,8 +23,8 @@ import android.util.Log;
 import com.firebase.ui.auth.ErrorCodes;
 import com.firebase.ui.auth.IdpResponse;
 import com.firebase.ui.auth.ResultCodes;
+import com.firebase.ui.auth.provider.ProviderUtils;
 import com.firebase.ui.auth.ui.BaseHelper;
-import com.firebase.ui.auth.ui.TaskFailureLogger;
 import com.firebase.ui.auth.ui.User;
 import com.firebase.ui.auth.ui.accountlink.WelcomeBackIdpPrompt;
 import com.firebase.ui.auth.ui.accountlink.WelcomeBackPasswordPrompt;
@@ -37,7 +37,6 @@ import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.EmailAuthProvider;
 import com.google.firebase.auth.FirebaseAuthUserCollisionException;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.auth.ProviderQueryResult;
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 public class CredentialSignInHandler implements OnCompleteListener<AuthResult> {
@@ -75,12 +74,9 @@ public class CredentialSignInHandler implements OnCompleteListener<AuthResult> {
                     mResponse);
         } else {
             if (task.getException() instanceof FirebaseAuthUserCollisionException) {
-                final String email = mResponse.getEmail();
+                String email = mResponse.getEmail();
                 if (email != null) {
-                    mHelper.getFirebaseAuth()
-                            .fetchProvidersForEmail(email)
-                            .addOnFailureListener(new TaskFailureLogger(
-                                    TAG, "Error fetching providers for email"))
+                    ProviderUtils.fetchTopProvider(mHelper.getFirebaseAuth(), email)
                             .addOnSuccessListener(new StartWelcomeBackFlow())
                             .addOnFailureListener(new OnFailureListener() {
                                 @Override
@@ -105,20 +101,22 @@ public class CredentialSignInHandler implements OnCompleteListener<AuthResult> {
         }
     }
 
-    private class StartWelcomeBackFlow implements OnSuccessListener<ProviderQueryResult> {
+    private class StartWelcomeBackFlow implements OnSuccessListener<String> {
         @Override
-        public void onSuccess(@NonNull ProviderQueryResult result) {
+        public void onSuccess(String provider) {
             mHelper.dismissDialog();
 
-            String provider = result.getProviders().get(0);
-            if (provider.equals(EmailAuthProvider.PROVIDER_ID)) {
+            if (provider == null) {
+                throw new IllegalStateException(
+                        "No provider even though we received a FirebaseAuthUserCollisionException");
+            } else if (provider.equals(EmailAuthProvider.PROVIDER_ID)) {
                 // Start email welcome back flow
                 mActivity.startActivityForResult(
                         WelcomeBackPasswordPrompt.createIntent(
                                 mActivity,
                                 mHelper.getFlowParams(),
-                                mResponse
-                        ), mAccountLinkRequestCode);
+                                mResponse),
+                        mAccountLinkRequestCode);
             } else {
                 // Start Idp welcome back flow
                 mActivity.startActivityForResult(
@@ -128,8 +126,8 @@ public class CredentialSignInHandler implements OnCompleteListener<AuthResult> {
                                 new User.Builder(mResponse.getEmail())
                                         .setProvider(provider)
                                         .build(),
-                                mResponse
-                        ), mAccountLinkRequestCode);
+                                mResponse),
+                        mAccountLinkRequestCode);
             }
         }
     }
